requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
beautifulsoup4>=4.12.0
chembl-webresource-client>=0.10.8
streamlit>=1.28.0
rdkit-pypi==2022.9.5

# Note: RDKit must be installed via conda:
# conda install -c conda-forge rdkit

